pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hakei"
version = "0.1.0"
description = "Electronic bench equipment visualization and control"
requires-python = ">=3.9"
dependencies = [
    "dearpygui>=1.11.1",
    "coloredlogs",
    "numpy",
    "psutil",
    "pydantic>=2.0",
    "pyvisa",
    "pyvisa-py",
    "pyyaml",
    "zeroconf",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "quartodoc",
]

[project.scripts]
hakei = "hakei.__main__:main"

[tool.setuptools.package-data]
hakei = ["fonts/*.ttf", "instruments/*.yaml"]

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["hakei"]
